server-agent/pyproject.toml
[project]
name = "pipeline-virtual-expert-agent"
version = "0.1.0"
description = "Virtual expert abnormality analysis Agent runtime"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.126.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "langgraph>=1.0.0",
  "langchain-core>=1.0.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.30.0",
  "redis>=5.2.0",
  "qdrant-client>=1.12.0",
  "httpx>=0.28.0",
  "python-jose[cryptography]>=3.3.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "ruff>=0.8.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
